A rough but at typed clients
This PR has **infrastructure** for typed clients. You either need to
write your own client classes, or use a library like refit.

What's here is the general pattern and plumbing. The intent is that a
library like refit would implement ITypedHttpClientFactory, OR you would
plug it using the overloads of AddTypedClient<T> that take a factory
function.

I'm introducing this now to try and get the discussion started about how
you might use this in MVC.